requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pypdf2==3.0.1
langchain==0.2.1
langchain-openai==0.1.7
langchain-community==0.2.1
faiss-cpu==1.8.0
sentence-transformers==2.7.0
transformers==4.41.2
torch==2.3.0
python-dotenv==1.0.1
